backend/requirements.txt
# Web Framework
Flask==3.0.2
flask-cors==4.0.0

#cryptography
cryptography==45.0.4

# Google API
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
google-cloud-speech==2.21.0

# OpenAI
openai==1.82.0

# Database
supabase==2.4.4

# Environment and Configuration
python-dotenv==1.0.1

# Date/Time Handling
pytz==2024.1

# JWT
PyJWT==2.8.0

# HTTP Client
requests==2.31.0

# Redis
redis==6.2.0
# Session Management
Flask-Session==0.7.0

# WSGI Server (for production)
gunicorn==21.2.0

cachelib>=0.9.0
